manifests: update manifests for CPU allocation
Internally the driver still allocates GPU but
with all the changes we have managed to enable
the user interface for CPU allocation.

The next set of changes will be made to update
the driver implmentation for allocation of CPU.

We expect some changes needed in Kubelet and/or
integration with custom NRI plugins for enabling
CPU allocation.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>