Ensure information passed to runtime after grpc communication
This is currently essentially a hard coded value but we know that the
gRPC communication works and the information is passed all the way
to the runtime.

Next Step: Ensure we make use of the cpuset request with pod and cpu
information and allocate cpus discovered by the kubelet plugin.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>